Nova Atualizacao

+++ wantfood-api/src/main/java/com/wantfood/aplication/infrastructure/repository/CozinhaRepositoryJPA.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -38,8 +39,12 @@ public class CozinhaRepositoryJPA implements CozinhaRepository{
 	//Removendo cozinha
 	@Transactional
 	@Override
-	public void remover(Cozinha cozinha) {
-		cozinha = porId(cozinha.getId());
+	public void remover(Long id) {
+		Cozinha cozinha = porId(id);
+		
+		if(cozinha == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
 		entityManager.remove(cozinha);
 	}
